chore: update folder path
