+++ finpro-api/prisma/schema.prisma
@@ -304,6 +304,8 @@ model CartItem {
   cart              Cart @relation(fields: [cartId], references: [id])
   productStock      ProductStock @relation(fields: [productId,storeId], references: [productId,storeId])
 
+  cartItemStatus   CartItemStatus @default(ACTIVE)
+
   @@map("cart_items")
 }
 
@@ -419,4 +421,10 @@ model PaymentProof {
   payment       Payment @relation(fields: [transactionId], references: [id])
 
   @@map("payment_proofs")
+}
+
+enum CartItemStatus {
+  ACTIVE
+  REMOVED
+  ORDERED
 }
